추천 기능 구현

+++ src/main/java/com/springstudy/backend/API/AI/Model/AIResponse.java
@@ -14,7 +14,7 @@ public record AIResponse(
 
         @Schema(description = "추천 영화 id")
         @NotNull
-        int movie_id,
+        Long movie_id,
 
         @Schema(description = "추천 이유")
         String aiResponse
